serialization: remove extra "hierarchy" from mediatypes
The fact that we are serializing is implied. This PR removes the extra
bits in the media type that only serve to confuse and waste bytes. The
new media types better reflect the reality for the usage of each type
and how they relate.

This must be followed up by a split of the serialization document, but
this change _should_ be applied first.

This change is backwards incompatible with previous versions of the OCI
specification. Since we are before 1.0, this is acceptable.

Signed-off-by: Stephen J Day <stephen.day@docker.com>