specs-go: clarify mediatypes
Due to the conflicting use of the `mediatType` field across
documents, and after discussion on
https://github.com/opencontainers/image-spec/pull/411,
this changeset removes the use of `mediaType` where it is used to refers
to a document's own type. Leaving only the use of `mediaType` for
descriptors, where it is used to describe the type of a referenced object.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>